moved show answer to its own component
